Fix compile on ESP32.
On ESP32, <avr/pgmspace.h> is located at <pgmspace.h>. Also on ESP32, the
compiler is more strict about integer conversions. It seems like a 'char'
cannot be used as an array index, probably because a 'char' is a distinct type
in C++, different from both 'signed char' and 'unsigned char'.